refactor event log skeleton to separate file
