+++ src/components/skeletons/index.ts
@@ -1,2 +1,3 @@
 export { default as Table } from "./table"
+export { default as EventLogs } from "./event-logs"
 export { default as PolicyFields } from "./policy-fields"
